requirements.txt
Flask==2.3.0
Flask-SocketIO==5.3.0
paho-mqtt==1.6.1
eventlet==0.33.1